+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.6] - 2025-12-03
+
+### Fixed
+- **[Hotfix]** Fix missing type `HumanoidPrefabInfo` in AvatarRegistrationWindow.cs
+  - Changed to use `PrefabInfo` from PackageAnalyzer
+
+---
+
 ## [0.0.5] - 2025-12-03
 
 ### Fixed